% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_respirometer.R
\name{extractor_o2meter}
\alias{extractor_o2meter}
\title{Extractor o2meter data}
\usage{
extractor_o2meter(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Convert all \code{.cvs} files in \code{o2meter} folder to \code{data.frame} and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database (and therefore make accessible for further analysis and forecasting)
from \code{.csv} files.
}
