% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processor_respirometer.R
\name{pre_processor_o2meter}
\alias{pre_processor_o2meter}
\title{Preprocessor o2meter data}
\usage{
pre_processor_o2meter(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Just copy from input to output
}
